+++ packages/bulma/src/Bulma.test.tsx
@@ -2,18 +2,15 @@ import { render, screen } from '@testing-library/react';
 import * as React from 'react';
 import { QueryBuilder, TestID } from 'react-querybuilder';
 import {
-  testActionElement,
   testNotToggle,
   testValueEditor,
   testValueSelector,
 } from 'react-querybuilder/genericTests/index';
-import { BulmaActionElement } from './BulmaActionElement';
 import { BulmaNotToggle } from './BulmaNotToggle';
 import { BulmaValueEditor } from './BulmaValueEditor';
 import { BulmaValueSelector } from './BulmaValueSelector';
 import { QueryBuilderBulma } from './index';
 
-testActionElement(BulmaActionElement);
 testNotToggle(BulmaNotToggle);
 testValueEditor(BulmaValueEditor);
 testValueSelector(BulmaValueSelector);